Changing to Controller format. Co-authored-by: aHjelmgaard <aHjelmgaard@users.noreply.github.com> Co-authored-by: Jakob <Kobo-coder@users.noreply.github.com>
