+++ src/webserver/Controllers/MainController.go
@@ -1,0 +1,8 @@
+package controllers
+
+import "github.com/antonPalmFolkmann/DevOps2022/models"
+
+type Response struct {
+	Data []models.User `json:"data"`
+	Message string `json:"message"`
+}
