+++ src/webserver/services/service.go
@@ -7,11 +7,6 @@ import (
 
 var dbconn *gorm.DB
 
-type Response struct {
-	Data []models.User `json:"data"`
-	Message string `json:"message"`
-}
-
 func SetDB(db *gorm.DB) {
 	dbconn = db
 	var user = models.GetUser()
